requirements.txt
celery==5.2.3
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flower==1.0.0
gunicorn==20.1.0
kombu==5.2.3
Pillow==9.0.1
psycopg2-binary==2.9.3
redis==4.1.4
stripe==2.65.0
sentry-sdk==1.5.8